my commit

+++ src/main/java/com/driver/services/BlogService.java
@@ -20,69 +20,54 @@ public class BlogService {
 
     @Autowired
     ImageService imageService1;
+
     @Autowired
     UserRepository userRepository1;
-    
-    
 
     public List<Blog> showBlogs(){
         //find all blogs
-        List<Blog> blogs = new ArrayList<>();
-        List<List<Blog>> listofblogs = new ArrayList<>();
-        for(User user : userRepository1.findAll()){
-            List<Blog> bloglist = user.getBlogList();
-            listofblogs.add(bloglist);
-        }
-        
-        for(List<Blog> x : listofblogs){
-            for(Blog y : x){
-                blogs.add(y);
-            }
-        }
-        return blogs;
+        return  blogRepository1.findAll();
+
     }
 
     public void createAndReturnBlog(Integer userId, String title, String content) {
         //create a blog at the current time
-        Blog blog = new Blog();
-        blog.setTitle(title);
-        blog.setContent(content);
-        User user = userRepository1.findById(userId).get();
-        blog.setUser(user);
+        Blog blog=new Blog(title,content,new Date());
+        blog.setUser(userRepository1.findById(userId).get());
+        User user=userRepository1.findById(userId).get();
+
 
         //updating the blog details
-        List<Blog> currentlistofBlogs = user.getBlogList();
-        currentlistofBlogs.add(blog);
-        user.setBlogList(currentlistofBlogs);
-
+        List<Blog> res=user.getBlogList();
+        res.add(blog);
+        user.setBlogList(res);
 
         //Updating the userInformation and changing its blogs
+        blogRepository1.save(blog);
         userRepository1.save(user);
-
     }
 
     public Blog findBlogById(int blogId){
-        //find a blog
         return blogRepository1.findById(blogId).get();
     }
 
     public void addImage(Integer blogId, String description, String dimensions){
-        Image image = new Image();
-        image.setDescription(description);
-        image.setDimensions(dimensions);
-        Blog blog = findBlogById(blogId);
+        //add an image to the blog after creating it
+        Blog blog=blogRepository1.findById(blogId).get();
+        Image image=imageService1.createAndReturn(blog,description,dimensions);
         image.setBlog(blog);
-
-        List<Image> imageList = blog.getImageList();
+        List<Image> imageList=blog.getImageList();
+        if(imageList==null) imageList=new ArrayList<>();
         imageList.add(image);
         blog.setImageList(imageList);
-
         blogRepository1.save(blog);
 
     }
 
     public void deleteBlog(int blogId){
         //delete blog and corresponding images
+        if(blogRepository1.findById(blogId).get()==null)
+            return;
         blogRepository1.deleteById(blogId);
     }
 }
